fix: documentation(undone)

+++ src/post/post.controller.ts
@@ -14,6 +14,8 @@ import { PostService } from './post.service';
 import { GetPostDto, PaginatedDto } from './dto/get-post.dto';
 import { CreatePostDto } from './dto/create-post.dto';
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
+import { ApiBearerAuth, ApiOkResponse } from '@nestjs/swagger';
+import { ApiTags } from '@nestjs/swagger';
 
 interface RequestWithUser extends Request {
   user: {
@@ -22,10 +24,13 @@ interface RequestWithUser extends Request {
 }
 
 @Controller('post')
+@ApiTags('Post')
 export class PostController {
   constructor(private readonly postsService: PostService) {}
 
   @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
+  @ApiOkResponse({ description: 'Post created' })
   @Post()
   addPost(@Body() post: CreatePostDto, @Req() req: RequestWithUser) {
     console.log(req.user);
@@ -33,6 +38,7 @@ export class PostController {
   }
 
   @Get()
+  @ApiOkResponse({ description: 'Get all posts.' })
   getAllPosts(
     @Query('page') page: string = '1',
     @Query('per') per: string = '10',
@@ -48,11 +54,14 @@ export class PostController {
   }
 
   @Get(':id')
+  @ApiOkResponse({ description: 'Get post by id.' })
   getPost(@Param('id') prodId: string) {
     return this.postsService.getSinglePost(prodId);
   }
 
   @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
+  @ApiOkResponse({ description: 'Update post by id.(only for author)' })
   @Patch(':id')
   async updatePost(
     @Param('id') prodId: string,
@@ -69,7 +78,16 @@ export class PostController {
   }
 
   @Delete(':id')
-  removePost(@Param('id') prodId: string): Promise<GetPostDto> {
-    return this.postsService.deletePost(prodId);
+  @ApiOkResponse({ description: 'Delete post by id.' })
+  async removePost(
+    @Param('id') prodId: string,
+    @Req() req: RequestWithUser,
+  ): Promise<GetPostDto> {
+    const author = await this.postsService.getAuthorId(prodId);
+    if (req.user.userId === author.authorId || req.user.userId === 'admin') {
+      return this.postsService.deletePost(prodId);
+    } else {
+      throw new Error('Unauthorized');
+    }
   }
 }
